Add middleware to support tracing-like functionalities
* Add TracingHook to logger for logging ctx
* Add traceLogger
* Add tracing middleware
* Support a logger to be used in tracing middleware
  - Log to file, Loglevel=trace
* Update API docs
* Split the existing middlewares into individual files
   - It could be helpful to clearly show callers in logs
* Increase default LOGFILE_MAXSIZE 1000 MB
