+++ Dockerfile
@@ -70,7 +70,7 @@ ENV TB_ROOT_PATH=/app \
     TB_DEFAULT_NAMESPACE=default \
     TB_DEFAULT_CREDENTIALHOLDER=admin \
     TB_LOGFILE_PATH=/app/log/tumblebug.log \
-    TB_LOGFILE_MAXSIZE=10 \
+    TB_LOGFILE_MAXSIZE=1000 \
     TB_LOGFILE_MAXBACKUPS=3 \
     TB_LOGFILE_MAXAGE=30 \
     TB_LOGFILE_COMPRESS=false \
