+++ src/api/rest/server/middlewares/responseBodyDump.go
@@ -3,120 +3,13 @@ package middlewares
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
-	"strings"
 	"time"
 
 	"github.com/cloud-barista/cb-tumblebug/src/core/common"
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
-
 	"github.com/rs/zerolog/log"
 )
-
-func Zerologger(skipPatterns [][]string) echo.MiddlewareFunc {
-	return middleware.RequestLoggerWithConfig(middleware.RequestLoggerConfig{
-		Skipper: func(c echo.Context) bool {
-			path := c.Request().URL.Path
-			query := c.Request().URL.RawQuery
-			for _, patterns := range skipPatterns {
-				isAllMatched := true
-				for _, pattern := range patterns {
-					if !strings.Contains(path+query, pattern) {
-						isAllMatched = false
-						break
-					}
-				}
-				if isAllMatched {
-					return true
-				}
-			}
-			return false
-		},
-		LogError:         true,
-		LogRequestID:     true,
-		LogRemoteIP:      true,
-		LogHost:          true,
-		LogMethod:        true,
-		LogURI:           true,
-		LogUserAgent:     false,
-		LogStatus:        true,
-		LogLatency:       true,
-		LogContentLength: true,
-		LogResponseSize:  true,
-		// HandleError:      true, // forwards error to the global error handler, so it can decide appropriate status code
-		LogValuesFunc: func(c echo.Context, v middleware.RequestLoggerValues) error {
-			if v.Error == nil {
-				log.Info().
-					Str("id", v.RequestID).
-					Str("client_ip", v.RemoteIP).
-					//Str("host", v.Host).
-					Str("method", v.Method).
-					Str("URI", v.URI).
-					//Str("user_agent", v.UserAgent).
-					Int("status", v.Status).
-					//Int64("latency", v.Latency.Nanoseconds()).
-					Str("latency_human", v.Latency.String()).
-					Str("bytes_in", v.ContentLength).
-					Int64("bytes_out", v.ResponseSize).
-					Msg("request")
-			} else {
-				log.Error().
-					Err(v.Error).
-					Str("id", v.RequestID).
-					Str("client_ip", v.RemoteIP).
-					// Str("host", v.Host).
-					Str("method", v.Method).
-					Str("URI", v.URI).
-					//Str("user_agent", v.UserAgent).
-					Int("status", v.Status).
-					// Int64("latency", v.Latency.Nanoseconds()).
-					Str("latency_human", v.Latency.String()).
-					Str("bytes_in", v.ContentLength).
-					Int64("bytes_out", v.ResponseSize).
-					Msg("request error")
-			}
-			return nil
-		},
-	})
-}
-
-func RequestIdAndDetailsIssuer(next echo.HandlerFunc) echo.HandlerFunc {
-	return func(c echo.Context) error {
-		// log.Debug().Msg("Start - Request ID middleware")
-
-		// Make X-Request-Id visible to all handlers
-		c.Response().Header().Set("Access-Control-Expose-Headers", echo.HeaderXRequestID)
-
-		// Get or generate Request ID
-		reqID := c.Request().Header.Get(echo.HeaderXRequestID)
-		if reqID == "" {
-			reqID = fmt.Sprintf("%d", time.Now().UnixNano())
-		}
-
-		// Set Request on the context
-		c.Set("RequestID", reqID)
-
-		//log.Trace().Msgf("(Request ID middleware) Request ID: %s", reqID)
-		if _, ok := common.RequestMap.Load(reqID); ok {
-			return fmt.Errorf("the X-Request-Id is already in use")
-		}
-
-		// Set "X-Request-Id" in response header
-		c.Response().Header().Set(echo.HeaderXRequestID, reqID)
-
-		details := common.RequestDetails{
-			StartTime:   time.Now(),
-			Status:      "Handling",
-			RequestInfo: common.ExtractRequestInfo(c.Request()),
-		}
-		common.RequestMap.Store(reqID, details)
-
-		// log.Debug().Msg("End - Request ID middleware")
-
-		return next(c)
-	}
-}
 
 func ResponseBodyDump() echo.MiddlewareFunc {
 	return middleware.BodyDumpWithConfig(middleware.BodyDumpConfig{
